+++ src/main/ArduinoTemperatureSensor.cpp
@@ -1,7 +1,7 @@
 #include <TemperatureSensor.h>
 
-float TemperatureSensor::readTemperature() {
+double TemperatureSensor::readTemperature() {
   int sensorVal = analogRead(pin);
-  float voltage = sensorVal * 5.0 / 1024;
+  double voltage = sensorVal * (5.0 / 1024);
   return (voltage - 0.5) * 100;
 }
